adding avatar view route

+++ src/modules/media/app/profile-avatar.service.ts
@@ -1,16 +1,16 @@
 import type e from "express";
-import { serveFile } from "../../../utils/nest.static.js";
+import { serveFile } from "../utils/nest.static.js";
 export const avatars_folder = ensuredJoinSync(PathsConfig.storage, "avatars", "base");
-import { ForbiddenException, NotFoundException } from "#/modules/errors/client-side/exceptions.js";
+import { ConflictException, ForbiddenException, NotFoundException } from "#/modules/errors/client-side/exceptions.js";
 import { join } from "node:path";
-import { create_avatar_hash_from_profile_id } from "../../../utils/crypto/hmac.js";
+import { create_avatar_hash_from_profile_id } from "../utils/crypto/hmac.js";
 import { existsSync } from "node:fs";
 import { Logger } from "log-it-colored";
 import { readFile, unlink, writeFile } from "node:fs/promises";
 import sharp from "sharp";
-import { MediaServiceUtils } from "../../../utils/methods.js";
+import { MediaServiceUtils } from "../utils/methods.js";
 import { PathsConfig } from "#/configs/paths.config.js";
-
+const tempProcessPath = ensuredJoinSync(PathsConfig.storage, "avatars", "temp");
 const avatar_image_width = 555 as const;
 const avatar_image_height = 555 as const;
 import { BadGatewayException, InternalServerErrorException } from "#/modules/errors/server-side/exceptions.js";
@@ -26,9 +26,9 @@ export const avatarService = new (class Avatar_Post_Service {
     avatar_delete = async ({ profile_id, avatar_url_hash }: Avatar_Delete_Service_Parameters) => {
         const actual_avatar_hash = create_avatar_hash_from_profile_id(profile_id);
         if (actual_avatar_hash !== avatar_url_hash) {
-            throw new ForbiddenException(["The provided avatar's Hash does not match"]);
+            throw new ForbiddenException(["Предоставленный хэш аватара не совпадает"]);
         }
-        const prod_path = join(PathsConfig.storage, "avatars", "base", `${avatar_url_hash}.webp`);
+        const prod_path = join(avatars_folder, `${avatar_url_hash}.webp`);
         if (!existsSync(prod_path)) {
             throw new NotFoundException(["Аватарка не найдена для удаления"]);
         }
@@ -36,9 +36,15 @@ export const avatarService = new (class Avatar_Post_Service {
     };
     avatar_set = async ({ profile_id, file }: avatar_upload_ServiceParameters) => {
         const avatar_hash = create_avatar_hash_from_profile_id(profile_id);
-        const prod_path = MediaServiceUtils.create_avatar_prod_path_FOR_UPLOAD(avatar_hash);
+        const prod_path = join(avatars_folder, `${avatar_hash}.webp`) as string;
+        if (existsSync(prod_path)) {
+            throw new ConflictException([
+                "Аватар с таким идентификатором профиля уже существует. Используйте другой идентификатор профиля или обновите существующий аватар.",
+            ]);
+        }
+
         const extname = MediaServiceUtils.get_correct_extname(file.mimetype);
-        const temp_path = join(PathsConfig.storage, "avatars", "temp", `${avatar_hash}.${extname}`);
+        const temp_path = join(tempProcessPath, `${avatar_hash}.${extname}`);
         await writeFile(temp_path, file.buffer);
         await image_sharp_process(temp_path, prod_path);
         return { avatar_hash } as const;
@@ -55,13 +61,9 @@ export const avatarService = new (class Avatar_Post_Service {
         await image_sharp_process(temp_path, prod_path);
         return { avatar_hash } as const;
     };
-    /** /avatar/:avatar_hash */
-    serveAvatarImage = async (req: e.Request, /**@Param("avatar_hash") */ avatar_hash: string, res: e.Response) => {
-        async function serve_avatar_files(req: e.Request, res: e.Response, avatar_hash: string) {
-            const filePath = join(avatars_folder, `${avatar_hash}.webp`);
-            return await serveFile(req, res, filePath);
-        }
-        return await serve_avatar_files(req, res, avatar_hash);
+    serveAvatarImage = async (user_cuid: string, req: e.Request, res: e.Response) => {
+        const filePath = join(avatars_folder, `${user_cuid}.webp`);
+        return await serveFile(req, res, filePath);
     };
 })();
